requirements.txt
python-telegram-bot==20.7
requests
feedparser
python-dotenv
beautifulsoup4
lxml
